# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: HerLoop API
  description: |
    HerLoop 她循环 — 女性社群闲置交易平台后端接口。
    所有接口返回 `{code, message, data}` 格式。分页接口使用 `PageResult<T>` 包装。
  version: 1.0.0
  contact:
    name: HerLoop Team
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
...
security:
  - BearerAuth: []
paths:
...
  /user/profile:
    get:
      tags: [User]
      summary: 获取当前用户信息
      responses:
        "200":
          description: 用户信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserVO"
    put:
      tags: [User]
      summary: 更新个人资料
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                avatar:
                  type: string
                  description: Emoji 头像
                wechat:
                  type: string
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserVO"
...
  /user/avatar:
    post:
      tags: [User]
      summary: 上传头像图片
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: 上传成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          url:
                            type: string
  /user/products:
    get:
      tags: [User]
      summary: 我发布的商品
      parameters:
        - $ref: "#/components/parameters/statusProduct"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: 商品列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageResultProductVO"
  /user/wants:
    get:
      tags: [User]
      summary: 我发布的求购
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: 求购列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageResultWantVO"
  /user/trades:
    get:
      tags: [User]
      summary: 我的交易记录
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [buy, sell]
          description: 买入/卖出筛选
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSizeDefault20"
      responses:
        "200":
          description: 交易记录
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageResultTradeVO"
...
  /user/points/log:
    get:
      tags: [User]
      summary: 积分流水
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSizeDefault20"
      responses:
        "200":
          description: 积分流水列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageResultPointsLogVO"

  # ── Product ─────────────────────────────────────────────
  /products:
    get:
      tags: [Product]
      summary: 商品列表（公开）
      security: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [全部, 服饰, 图书, 美妆, 运动, 生活]
        - name: tradeMode
          in: query
          schema:
            type: string
            enum: [cash, points, both]
        - name: status
          in: query
          schema:
            type: string
            enum: [on, reserved, sold]
        - name: keyword
          in: query
          schema:
            type: string
          description: 关键词搜索（FULLTEXT on title+description）
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: 商品列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageResultProductVO"
    post:
      tags: [Product]
      summary: 发布商品
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreateRequest"
      responses:
        "200":
          description: 发布成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ProductVO"
...
  /wants:
    get:
      tags: [Want]
      summary: 求购列表（公开）
      security: []
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: 求购列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PageResultWantVO"
    post:
      tags: [Want]
      summary: 发布求购
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WantCreateRequest"
      responses:
        "200":
          description: 发布成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/WantVO"
  /wants/{id}:
    get:
      tags: [Want]
      summary: 求购详情（公开）
      security: []
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: 求购详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/WantVO"
  /wants/{id}/close:
    patch:
      tags: [Want]
      summary: 关闭求购
      parameters:
        - $ref: "#/components/parameters/idPath"
      responses:
        "200":
          description: 关闭成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /wants/{id}/images:
    post:
      tags: [Want]
      summary: 上传求购图片（最多 5 张）
      parameters:
        - $ref: "#/components/parameters/idPath"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 5
      responses:
        "200":
          description: 上传成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: string

  # ── Trade ───────────────────────────────────────────────
  /trades:
    post:
      tags: [Trade]
      summary: 创建交易
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeCreateRequest"
      responses:
        "200":
          description: 交易创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Result"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TradeVO"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    idPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 10
    pageSizeDefault20:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 20
    statusProduct:
      name: status
      in: query
      schema:
        type: string
        enum: [on, reserved, sold]
  schemas:
...
    Result:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          nullable: true

    # ── Auth ──
...
    UserVO:
      type: object
      properties:
        id:
          type: integer
        nickname:
          type: string
        avatar:
          type: string
        wechat:
          type: string
        email:
          type: string
        points:
          type: integer
        verified:
          type: boolean
        inviteCode:
          type: string
        productCount:
          type: integer
        tradeCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    # ── Product ──
    ProductCreateRequest:
      type: object
      required: [title, category, condition, description, tradeMode]
      properties:
        title:
          type: string
        category:
          type: string
          enum: [服饰, 图书, 美妆, 运动, 生活]
        condition:
          type: string
          example: 95新
        description:
          type: string
        cashPrice:
          type: number
        pointsPrice:
          type: integer
        tradeMode:
          type: string
          enum: [cash, points, both]
        icon:
          type: string
        wechat:
          type: string
    ProductVO:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        condition:
          type: string
        description:
          type: string
        cashPrice:
          type: number
        pointsPrice:
          type: integer
        tradeMode:
          type: string
          enum: [cash, points, both]
        status:
          type: string
          enum: [on, reserved, sold]
        icon:
          type: string
        imageUrls:
          type: string
          description: 逗号分隔的图片 URL
        createdAt:
          type: string
          format: date-time
        seller:
          $ref: "#/components/schemas/SellerBrief"
    SellerBrief:
      type: object
      properties:
        id:
          type: integer
        nickname:
          type: string
        avatar:
          type: string
        verified:
          type: boolean

    # ── Want ──
    WantCreateRequest:
      type: object
      required: [title, budget, description]
      properties:
        title:
          type: string
        budget:
          type: string
          example: ¥300以内
        description:
          type: string
        icon:
          type: string
    WantVO:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        budget:
          type: string
        description:
          type: string
        icon:
          type: string
        imageUrls:
          type: string
        status:
          type: string
          enum: [open, closed]
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/SellerBrief"

    # ── Trade ──
    TradeCreateRequest:
      type: object
      required: [productId]
      properties:
        productId:
          type: integer
    TradeVO:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/ProductVO"
        type:
          type: string
          enum: [buy, sell]
        status:
          type: string
        pointsPaid:
          type: integer
          description: 本次交易消耗的积分
        tradeDate:
          type: string
          format: date-time

    # ── Points ──
    PointsLogVO:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        amount:
          type: string
          example: "+60"
        expireAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    # ── Notification ──
...
    PageResultProductVO:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/ProductVO"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    PageResultWantVO:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/WantVO"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    PageResultTradeVO:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/TradeVO"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    PageResultPointsLogVO:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: "#/components/schemas/PointsLogVO"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
...